[rl] Move Verifiers integration into example
